interface 定数を除去
