+++ src/main/java/quantumshogi/player/Player.kt
@@ -1,0 +1,5 @@
+package quantumshogi.player
+
+enum class Player(val direction: Int) {
+    P1(1), P2(-1)
+}
